Move non-reused pipeline pieces into the training script, some pipeline refactoring
